requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dateutils==0.6.12
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.44.0
django-cleanup==5.1.0
django-decouple==2.1
django-filter==2.4.0
django-social-share==2.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
imutils==0.5.4
nsetools==1.0.11
numpy==1.20.3
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2019.1
requests==2.25.1
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.3.0
update-checker==0.18.0
urllib3==1.26.4
whitenoise==5.2.0
